+++ CHANGES.rst
@@ -16,9 +16,11 @@ Changelog
 
 
 **Fixed**
+
 - #991 New client contacts do not have access to their own AR Templates
 - #996 Hide checkbox labels on category expansion
 - #990 Fix client analysisspecs view
+- #899 Secondary AR changes received date of sample
 
 
 **Security**